requirements.txt
open3d 
memory_profiler
trimesh 
pytorch 

